deduplicate all of the gl1/gl3 texture code
